More bagit stuff (untested)
