+++ demo/src/main/edu/georgetown/library/fileAnalyzer/filetest/VerifyAPTrustBag.java
@@ -1,15 +1,8 @@
 package edu.georgetown.library.fileAnalyzer.filetest;
 
-import gov.loc.repository.bagit.Bag;
-import gov.loc.repository.bagit.BagFactory;
-import gov.loc.repository.bagit.utilities.SimpleResult;
 import gov.nara.nwts.ftapp.FTDriver;
-import gov.nara.nwts.ftapp.filetest.DefaultFileTest;
 import gov.nara.nwts.ftapp.stats.Stats;
-import gov.nara.nwts.ftapp.stats.StatsGenerator;
-import gov.nara.nwts.ftapp.stats.StatsItem;
 import gov.nara.nwts.ftapp.stats.StatsItemConfig;
-import gov.nara.nwts.ftapp.stats.StatsItemEnum;
 
 import java.io.File;
 
@@ -18,38 +11,7 @@ import java.io.File;
  * @author TBrady
  *
  */
-class VerifyAPTrustBag extends DefaultFileTest { 
-    public enum STAT {
-    	VALID,
-        INVALID, 
-        ERROR,
-        NA
-    }
-    
-    static enum BagStatsItems implements StatsItemEnum {
-        Key(StatsItem.makeStringStatsItem("Bag Path", 200)),
-        Stat(StatsItem.makeEnumStatsItem(STAT.class, "Bag Status")),
-        Count(StatsItem.makeIntStatsItem("Item Count")),
-        Message(StatsItem.makeStringStatsItem("Message",400)),
-        ;
-        StatsItem si;
-        BagStatsItems(StatsItem si) {this.si=si;}
-        public StatsItem si() {return si;}
-    }
-
-    public static enum Generator implements StatsGenerator {
-        INSTANCE;
-        class BagStats extends Stats {
-            public BagStats(String key) {
-                super(details, key);
-            }
-
-        }
-        public BagStats create(String key) {return new BagStats(key);}
-    }
-    public static StatsItemConfig details = StatsItemConfig.create(BagStatsItems.class);
-
-    long counter = 1000000;
+class VerifyAPTrustBag extends VerifyBag { 
     public VerifyAPTrustBag(FTDriver dt) {
         super(dt);
     }
@@ -63,23 +25,6 @@ class VerifyAPTrustBag extends DefaultFileTest {
     
     public String getShortName(){return "Ver APT Dir";}
 
-    
-    public Object fileTest(File f) {
-        Stats s = getStats(f);
-        BagFactory bf = new BagFactory();
-        Bag bag = bf.createBag(f);
-        s.setVal(BagStatsItems.Count, bag.getPayload().size());
-        SimpleResult result = bag.verifyValid();
-        if (result.isSuccess()) {
-            s.setVal(BagStatsItems.Stat, STAT.VALID);
-        } else {
-            s.setVal(BagStatsItems.Stat, STAT.INVALID);
-            for(String m: result.getMessages()) {
-                s.appendVal(BagStatsItems.Message, m +" ");             
-            }
-        }
-        return s.getVal(BagStatsItems.Count);
-    }
     public Stats createStats(String key){ 
         return Generator.INSTANCE.create(key);
     }
@@ -94,19 +39,8 @@ class VerifyAPTrustBag extends DefaultFileTest {
         		+ "'assume directory names do not contain periods' on the advanced tab.";
     }
     
-    @Override public boolean isTestDirectory(File f) {
-    	return hasDescendant(f, "bagit.txt");
-    }
-    @Override public boolean processRoot() {
-        return true;
-    }
-
-    @Override public boolean isTestFiles() {
-        return false; 
-    }
-
     @Override public boolean isTestable(File f) {
-        return (new File(f, "bagit.txt")).exists();
+        return hasBagFile(f);
     }
