+++ demo/src/main/edu/georgetown/library/fileAnalyzer/filetest/VerifyBag.java
@@ -5,6 +5,7 @@ import gov.loc.repository.bagit.BagFactory;
 import gov.loc.repository.bagit.utilities.SimpleResult;
 import gov.nara.nwts.ftapp.FTDriver;
 import gov.nara.nwts.ftapp.filetest.DefaultFileTest;
+import gov.nara.nwts.ftapp.ftprop.FTPropEnum;
 import gov.nara.nwts.ftapp.stats.Stats;
 import gov.nara.nwts.ftapp.stats.StatsGenerator;
 import gov.nara.nwts.ftapp.stats.StatsItem;
@@ -12,6 +13,8 @@ import gov.nara.nwts.ftapp.stats.StatsItemConfig;
 import gov.nara.nwts.ftapp.stats.StatsItemEnum;
 
 import java.io.File;
+
+import edu.georgetown.library.fileAnalyzer.BAG_TYPE;
 
 /**
  * Extract all metadata fields from a TIF or JPG using categorized tag defintions.
@@ -22,10 +25,14 @@ class VerifyBag extends DefaultFileTest {
     public enum STAT {
         VALID,
         INVALID, 
-        ERROR
+        ERROR,
+        NA
     }
     
-    private static enum BagStatsItems implements StatsItemEnum {
+    private FTPropEnum pBagType = new FTPropEnum(dt, this.getClass().getSimpleName(),  CreateBag.P_BAGTYPE, CreateBag.P_BAGTYPE,
+            "Type of bag to create", BAG_TYPE.values(), BAG_TYPE.DIRECTORY);
+
+    static enum BagStatsItems implements StatsItemEnum {
         Key(StatsItem.makeStringStatsItem("Bag Path", 200)),
         Stat(StatsItem.makeEnumStatsItem(STAT.class, "Bag Status")),
         Count(StatsItem.makeIntStatsItem("Item Count")),
@@ -51,32 +58,37 @@ class VerifyBag extends DefaultFileTest {
     long counter = 1000000;
     public VerifyBag(FTDriver dt) {
         super(dt);
+        ftprops.add(pBagType);
     }
 
     public String toString() {
-        return "Verify Bag";
+        return "Verify Bag - Dir";
     }
     public String getKey(File f) {
         return this.getRelPath(f);
     }
     
-    public String getShortName(){return "Bag";}
+    public String getShortName(){return "Ver Bag";}
 
     
     public Object fileTest(File f) {
         Stats s = getStats(f);
         BagFactory bf = new BagFactory();
-        Bag bag = bf.createBag(f);
-        s.setVal(BagStatsItems.Count, bag.getPayload().size());
-        SimpleResult result = bag.verifyValid();
-        if (result.isSuccess()) {
-            s.setVal(BagStatsItems.Stat, STAT.VALID);
-        } else {
-            s.setVal(BagStatsItems.Stat, STAT.INVALID);
-            for(String m: result.getMessages()) {
-                s.appendVal(BagStatsItems.Message, m +" ");             
-            }
-        }
+        try (Bag bag = bf.createBag(f);) {
+			s.setVal(BagStatsItems.Count, bag.getPayload().size());
+			SimpleResult result = bag.verifyValid();
+			if (result.isSuccess()) {
+			    s.setVal(BagStatsItems.Stat, STAT.VALID);
+			} else {
+			    s.setVal(BagStatsItems.Stat, STAT.INVALID);
+			    for(String m: result.getMessages()) {
+			        s.appendVal(BagStatsItems.Message, m +" ");             
+			    }
+			}
+		} catch (Exception e) {
+	        s.setVal(BagStatsItems.Message, e.getMessage());
+		    s.setVal(BagStatsItems.Stat, STAT.NA);
+		}
         return s.getVal(BagStatsItems.Count);
     }
     public Stats createStats(String key){ 
@@ -90,20 +102,22 @@ class VerifyBag extends DefaultFileTest {
         return "This rule will scan for directories with names ending with '_bag'.  BagIt verifyvalid will be run on the bag.";
     }
     
-    @Override public boolean processRoot() {
-        return true;
+    @Override public boolean isTestDirectory(File f) {
+    	return hasDescendant(f, "bagit.txt");
     }
-    @Override public boolean isTestDirectory() {
+    @Override public boolean processRoot() {
         return true;
     }
 
     @Override public boolean isTestFiles() {
         return false; 
     }
-    
+
+    public boolean hasBagFile(File f) {
+        return (new File(f, "bagit.txt")).exists();
+    }
     @Override public boolean isTestable(File f) {
-        //if (!f.getParentFile().equals(dt.root)) return false;
-        return (f.getName().endsWith("_bag"));
+        return hasBagFile(f) && f.getName().endsWith("_bag");
     }
 
 }
