+++ demo/src/main/edu/georgetown/library/fileAnalyzer/filetest/VerifyBagZip.java
@@ -1,8 +1,5 @@
 package edu.georgetown.library.fileAnalyzer.filetest;
 
-import gov.loc.repository.bagit.Bag;
-import gov.loc.repository.bagit.BagFactory;
-import gov.loc.repository.bagit.utilities.SimpleResult;
 import gov.nara.nwts.ftapp.FTDriver;
 import gov.nara.nwts.ftapp.filter.ZipFilter;
 import gov.nara.nwts.ftapp.stats.Stats;
@@ -15,18 +12,17 @@ import java.io.File;
  * @author TBrady
  *
  */
-class VerifyAPTrustBagZip extends VerifyBagZip { 
+class VerifyBagZip extends VerifyBag { 
     public static StatsItemConfig details = StatsItemConfig.create(BagStatsItems.class);
 
-    public VerifyAPTrustBagZip(FTDriver dt) {
+    public VerifyBagZip(FTDriver dt) {
         super(dt);
     }
 
     public String toString() {
-        return "Verify APTrust Bag - Zip";
+        return "Verify Bag - Zip";
     }
     public String getShortName(){return "Ver APT Zip";}
-
     
     public Stats createStats(String key){ 
         return Generator.INSTANCE.create(key);
@@ -36,7 +32,27 @@ class VerifyAPTrustBagZip extends VerifyBagZip {
     }
 
     public String getDescription() {
-        return "This rule will validate the contents of an APTrust bag zip file";
+        return "This rule will validate the contents of a bag zip file";
     }
+    
+    @Override public boolean isTestDirectory() {
+    	return false;
+    }
+    @Override public boolean processRoot() {
+        return false;
+    }
+
+    @Override public boolean isTestFiles() {
+        return true; 
+    }
+
+    @Override public boolean isTestable(File f) {
+    	//TODO validate name
+        return true;
+    }
+
+	public void initFilters() {
+		filters.add(new ZipFilter());
+	}
     
 }
